requirements.txt
alembic
annotated-doc
annotated-types
anyio
async-timeout
asyncpg
certifi
click
dnspython
email-validator
exceptiongroup
fastapi
fastapi-cli
fastapi-cloud-cli
fastar
greenlet
h11
httpcore
httptools
httpx
idna
itsdangerous
Jinja2
Mako
markdown-it-py
MarkupSafe
mdurl
orjson
passlib
psycopg2-binary
pydantic
pydantic-extra-types
pydantic-settings
pydantic_core
Pygments
PyJWT
python-dotenv
python-multipart
PyYAML
rich
rich-toolkit
rignore
sentry-sdk
shellingham
SQLAlchemy
starlette
tomli
typer
typing-inspection
typing_extensions
ujson
urllib3
uvicorn
uvloop
watchfiles
websockets
